Support running podman and docker-compose from toolbox container
Install 'podman-remote' in dev container and alias back to 'podman'
so that podman can be run within the container and it will map through
to the host installation making it feel more native.

Install 'docker-compose' so that service containers can be setup.

Pre-configure 'podman' to default to docker registry for images and
avoid remapping user ids when conatainers are not run as root. This
helps the setup feel much more like docker itself without needing
root so easier to run on Fedora atomic distros.